requirements.txt
appnope==0.1.2
asgiref==3.4.1
backcall==0.2.0
decorator==5.1.0
Django==3.2.9
django-filter==21.1
djangorestframework==3.12.4
factory-boy==3.2.1
Faker==9.9.0
ipython==7.30.0
jedi==0.18.1
matplotlib-inline==0.1.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.23
ptyprocess==0.7.0
Pygments==2.10.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
traitlets==5.1.1
wcwidth==0.2.5
